requirements.txt
fastapi[all]
uvicorn
Pillow
pytesseract
pdfplumber